scrapy/requirements.txt
# Editor
pylint
black

# Local requirements
Scrapy>=2.6.2
defusedxml>=0.7.1
scrapydo>=0.2.2
typesense>=0.14.0

# Amazon AWS
# boto3>=1.21.23

# Microsoft Azure

# Google Cloud
functions-framework
google-cloud-firestore
google-cloud-tasks